Time: 27 ms (23.81%), Space: 45.3 MB (27.3%) - LeetHub

+++ 0907-koko-eating-bananas/0907-koko-eating-bananas.java
@@ -33,7 +33,7 @@ class Solution
         int sum=0;
         for(int i=0;i<piles.length;i++)
         {
-            sum+= Math.ceil((double) piles[i]/(double)mid);
+            sum+= Math.ceil((double) piles[i]/mid);
         }
 
         return sum;
